requirements.txt
Django==5.1.12
psycopg[binary]==3.2.3
django-storages==1.14.4
boto3==1.35.28
djangorestframework==3.15.2
